+++ source/rafcon/gui/helpers/state.py
@@ -528,12 +528,14 @@ def substitute_state(target_state_m, state_m_to_insert, as_template=False):
                                                    action_parent_m=action_parent_m,
                                                    affected_models=[old_state_m, ], after=False,
                                                    kwargs={'state_id': state_id, 'state': state_to_insert}))
-    related_transitions, related_data_flows = action_parent_m.state.related_linkage_state(state_id)
+    related_transitions, related_data_flows = action_parent_m.state.get_connections_for_state(state_id)
     tmp_meta_data['state'] = old_state_m.meta
     # print("old state meta", old_state_m.meta)
-    for t in set(related_transitions['external']['ingoing'] + related_transitions['external']['outgoing']):
+    external_t = related_transitions['external']
+    for t in external_t['ingoing'] + external_t['outgoing'] + external_t['self']:
         tmp_meta_data['transitions'][t.transition_id] = action_parent_m.get_transition_m(t.transition_id).meta
-    for df in set(related_data_flows['external']['ingoing'] + related_data_flows['external']['outgoing']):
+    external_df = related_data_flows['external']
+    for df in external_df['ingoing'] + external_df['outgoing'] + external_df['self']:
         tmp_meta_data['data_flows'][df.data_flow_id] = action_parent_m.get_data_flow_m(df.data_flow_id).meta
     action_parent_m.substitute_state.__func__.tmp_meta_data_storage = tmp_meta_data
     action_parent_m.substitute_state.__func__.old_state_m = old_state_m
@@ -638,7 +640,7 @@ def group_states_and_scoped_variables(state_m_list, sv_m_list):
     tmp_models_dict = {'transitions': {}, 'data_flows': {}, 'states': {}, 'scoped_variables': {}, 'state': None,
                        'input_data_ports': {}, 'output_data_ports': {}}
     related_transitions, related_data_flows = \
-        action_parent_m.state.related_linkage_states_and_scoped_variables(state_ids, sv_ids)
+        action_parent_m.state.get_connections_for_state_and_scoped_variables(state_ids, sv_ids)
     for state_id in state_ids:
         tmp_models_dict['states'][state_id] = action_parent_m.states[state_id]
     for sv_id in sv_ids:
@@ -649,11 +651,11 @@ def group_states_and_scoped_variables(state_m_list, sv_m_list):
         tmp_models_dict['data_flows'][df.data_flow_id] = action_parent_m.get_data_flow_m(df.data_flow_id)
 
     affected_models = []
-    for elemets_dict in tmp_models_dict.values():
-        if isinstance(elemets_dict, dict):
-            affected_models.extend(elemets_dict.values())
-        elif isinstance(elemets_dict, AbstractStateModel):
-            affected_models.extend(elemets_dict)
+    for elements_dict in tmp_models_dict.values():
+        if isinstance(elements_dict, dict):
+            affected_models.extend(elements_dict.values())
+        elif isinstance(elements_dict, AbstractStateModel):
+            affected_models.extend(elements_dict)
 
     # print("EMIT-BEFORE ON ACTION PARENT")
     action_parent_m.action_signal.emit(ActionSignalMsg(action='group_states', origin='model',
@@ -718,7 +720,7 @@ def ungroup_state(state_m):
     tmp_models_dict = {'transitions': {}, 'data_flows': {}, 'states': {}, 'scoped_variables': {}, 'state': None,
                        'input_data_ports': {}, 'output_data_ports': {}}
 
-    related_transitions, related_data_flows = action_parent_m.state.related_linkage_state(state_id)
+    related_transitions, related_data_flows = action_parent_m.state.get_connections_for_state(state_id)
     tmp_models_dict['state'] = action_parent_m.states[state_id]
     for s_id, s_m in action_parent_m.states[state_id].states.items():
         tmp_models_dict['states'][s_id] = s_m