+++ doc/changelog.rst
@@ -18,6 +18,8 @@ Next Release
 
 
 - Bug Fixes:
+
+  - :issue_ghe:`726` State with self-transition cannot be substituted
   - fixes bug of "locked" global variable during multithreading access
 
 - Changes: